+++ firmware/hdl/payload/gtl/bx_pipeline.vhd
@@ -2,6 +2,8 @@
 -- Pipeline for +/-2 bx data.
 
 -- Version-history:
+-- HB 2023-06-23: cicada on link 11 (removed zdc10g on link 11).
+-- HB 2022-10-10: updated for ZDC (on link 11).
 -- HB 2021-06-10: updated for hadronic shower trigger (mus).
 -- HB 2021-05-21: first design.
 
@@ -29,8 +31,6 @@ architecture rtl of bx_pipeline is
 
     type mus_tmp_array is array (0 to BX_PIPELINE_STAGES-1) of std_logic_vector(NR_MUS_BITS-1 downto 0);
     signal mus_tmp, mus_tmp_del : mus_tmp_array;
-
-    signal mus0, mus1, musoot0, musoot1 : mus_bit_array;
 
 begin
 
@@ -103,6 +103,7 @@ begin
         mus_tmp(i)(1) <= data_tmp(i).mu(MUON_OBJ_MUS1)(MUS_BIT);
         mus_tmp(i)(2) <= data_tmp(i).mu(MUON_OBJ_MUSOOT0)(MUS_BIT);
         mus_tmp(i)(3) <= data_tmp(i).mu(MUON_OBJ_MUSOOT1)(MUS_BIT);
+        mus_tmp(i)(4) <= data_tmp(i).mu(MUON_OBJ_MUS2)(MUS_BIT);
 
         mus_pipe_i: entity work.delay_pipeline
             generic map(
@@ -117,6 +118,7 @@ begin
         bx_data.mus1(i) <= mus_tmp_del(i)(1);
         bx_data.musoot0(i) <= mus_tmp_del(i)(2);
         bx_data.musoot1(i) <= mus_tmp_del(i)(3);
+        bx_data.mus2(i) <= mus_tmp_del(i)(4);
 
     end generate bx_l;