+++ firmware/hdl/payload/gtl/comb_conditions.vhd
@@ -42,6 +42,17 @@ entity comb_conditions is
         eta_w4_lower_limits_obj1: common_templates_array := (others => (others => '0'));
         eta_w5_upper_limits_obj1: common_templates_array := (others => (others => '0'));
         eta_w5_lower_limits_obj1: common_templates_array := (others => (others => '0'));
+        nr_idx_windows_obj1: common_templates_natural_array := (others => 0);
+        idx_w1_upper_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w1_lower_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w2_upper_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w2_lower_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w3_upper_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w3_lower_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w4_upper_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w4_lower_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w5_upper_limits_obj1: common_templates_array := (others => (others => '0'));
+        idx_w5_lower_limits_obj1: common_templates_array := (others => (others => '0'));
         nr_phi_windows_obj1: common_templates_natural_array := (others => 0);
         phi_w1_upper_limits_obj1: common_templates_array := (others => (others => '0'));
         phi_w1_lower_limits_obj1: common_templates_array := (others => (others => '0'));
@@ -80,6 +91,17 @@ entity comb_conditions is
         eta_w4_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
         eta_w5_upper_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
         eta_w5_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        nr_idx_windows_obj2: natural := 0;
+        idx_w1_upper_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w1_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w2_upper_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w2_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w3_upper_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w3_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w4_upper_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w4_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w5_upper_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
+        idx_w5_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
         nr_phi_windows_obj2: natural := 0;
         phi_w1_upper_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
         phi_w1_lower_limit_obj2: std_logic_vector(MAX_TEMPLATES_BITS-1 downto 0) := (others => '0');
@@ -326,6 +348,12 @@ begin
                 eta_w3_upper_limits_obj1, eta_w3_lower_limits_obj1,
                 eta_w4_upper_limits_obj1, eta_w4_lower_limits_obj1,
                 eta_w5_upper_limits_obj1, eta_w5_lower_limits_obj1,
+                nr_idx_windows_obj1,
+                idx_w1_upper_limits_obj1, idx_w1_lower_limits_obj1,
+                idx_w2_upper_limits_obj1, idx_w2_lower_limits_obj1,
+                idx_w3_upper_limits_obj1, idx_w3_lower_limits_obj1,
+                idx_w4_upper_limits_obj1, idx_w4_lower_limits_obj1,
+                idx_w5_upper_limits_obj1, idx_w5_lower_limits_obj1,
                 nr_phi_windows_obj1,
                 phi_w1_upper_limits_obj1, phi_w1_lower_limits_obj1,
                 phi_w2_upper_limits_obj1, phi_w2_lower_limits_obj1,