+++ scripts/ignored_algos.py
@@ -1,4 +1,4 @@
 IGNORED_ALGOS = [
     'L1_FirstBunchInTrain',
-    'L1_SecondBunchInTrain',
+    'L1_SecondBunchInTrain'
 ]